requirements.txt
anthropic==0.42.0
google-generativeai==0.8.3
mcp==1.1.2
PyMuPDF==1.25.3
python-dotenv==1.0.1
openai>=1.0.0
